jarvis/requirements.txt
flask
flask-socketio
openai
psutil
requests
Pillow
